int64 to int

+++ internal/storage/tribonacciInMemory/get.go
@@ -2,7 +2,7 @@ package tribonacciInMemory
 
 import "github.com/MKuchum/tada-testing/models"
 
-func (s *Storage) Get(signature []float32, n int64) ([]float32, error) {
+func (s *Storage) Get(signature []float32, n int) ([]float32, error) {
 	if len(signature) != 3 {
 		return nil, models.WrongSignatureLenErr
 	}
@@ -12,7 +12,7 @@ func (s *Storage) Get(signature []float32, n int64) ([]float32, error) {
 
 	key := s.genKey(signature)
 	if v, ok := s.m[key]; ok {
-		if len(v) >= int(n) {
+		if len(v) >= n {
 			return v[:n], nil
 		} else {
 			return v, nil
